Jsonb refactor

+++ tfc_web/csn/views.py
@@ -1,57 +1,54 @@
 import json
 import requests
 from datetime import datetime, timedelta
+from collections import namedtuple
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.core.exceptions import ValidationError
-from django.shortcuts import render, redirect, get_object_or_404
+from django.http import HttpResponseNotFound
+from django.shortcuts import render, redirect
 from csn.everynet_api import everynet_add_device, everynet_remove_device
-from csn.forms import LWDeviceForm, LWApplicationForm
-from csn.models import LWDevice, LWApplication, LOGGER
+from csn.forms import LWDeviceForm, LWApplicationForm, LWDeviceFormExtended
+from csn.models import LOGGER, Destination, Sensor
 
 
 @login_required
 def devices(request):
     return render(request, 'csn/devices.html', {
-        'devices': LWDevice.objects.filter(user=request.user),
-        'app_choices': LWApplication.objects.filter(user=request.user)
+        'devices': Sensor.get_all_lorawan_with_lwapps(user_id=request.user.id),
+        'app_choices': Destination.get_all_everynet_jsonrpc(user_id=request.user.id)
     })
 
 
 @login_required
 def device(request, device_id):
-    lwdevice = get_object_or_404(LWDevice, user=request.user, id=device_id)
+    lwdevice = Sensor.get_lorawan(sensor_id=device_id, user_id=request.user.id)
     return render(request, 'csn/device.html', {
         'device': lwdevice,
-    })
+    }) if lwdevice else HttpResponseNotFound()
 
 
 @login_required
 def new_device(request):
     lwdevice_form = LWDeviceForm(user=request.user)
     if request.method == "POST":
-        lwdevice_form = LWDeviceForm(request.POST, user=request.user)
+        lwdevice_form = LWDeviceFormExtended(request.POST, user=request.user)
         try:
             if lwdevice_form.is_valid():
-                lwdevice = lwdevice_form.save(commit=False)
-                lwdevice.activation_type = request.POST['activation_type']
-                if lwdevice.activation_type == "abp":
-                    lwdevice.nwkskey = request.POST['nwkskey']
-                    lwdevice.appskey = request.POST['appskey']
-                    lwdevice.dev_addr = request.POST['dev_addr']
-                elif lwdevice.activation_type == "otaa":
-                    lwdevice.app_key = request.POST['app_key']
-                else:
-                    raise ValidationError("Activation type not supported")
-                lwdevice.full_clean(exclude=['tfcserver_sensor_id'])
-                if everynet_add_device(lwdevice):
-                    lwdevice.save()
+                lwdevice = lwdevice_form.cleaned_data
+                lwdevice_named = namedtuple("LWDevice", lwdevice.keys())(*lwdevice.values())
+                if everynet_add_device(lwdevice_named):
+                    Sensor.insert_lorawan(info=lwdevice)
                     return redirect('csn_devices')
                 else:
                     lwdevice_form.add_error(field=None, error=lwdevice.error_message)
         except Exception as e:
             lwdevice_form.add_error(field=None, error=str(e))
+        lwdevice_form.fields.pop('activation_type')
+        lwdevice_form.fields.pop('nwkskey')
+        lwdevice_form.fields.pop('appskey')
+        lwdevice_form.fields.pop('dev_addr')
+        lwdevice_form.fields.pop('app_key')
     return render(request, 'csn/new_device.html', {
         'form': lwdevice_form
     })
@@ -60,8 +57,11 @@ def new_device(request):
 @login_required
 def delete_device(request):
     if request.method == "POST":
-        lwdevice = get_object_or_404(LWDevice, user=request.user, dev_eui=request.POST['dev_eui'])
+        lwdevice = Sensor.get_lorawan(sensor_id=request.POST['sensor_id'], user_id=request.user.id)
+        if not lwdevice:
+            return HttpResponseNotFound()
         if everynet_remove_device(lwdevice):
+            # TODO change this to Sensor.delete_lorawan?
             lwdevice.delete()
             messages.info(request, "Device deleted")
         else:
@@ -72,9 +72,12 @@ def delete_device(request):
 @login_required
 def change_device_app(request):
     if request.method == "POST":
-        lwdevice = get_object_or_404(LWDevice, user=request.user, id=request.POST['dev_id'])
-        lwapp = get_object_or_404(LWApplication, user=request.user, id=request.POST['app_id'])
-        lwdevice.lw_application = lwapp
+        lwdevice = Sensor.get_lorawan(sensor_id=request.POST['sensor_id'], user_id=request.user.id)
+        lwapp = Destination.get_everynet_jsonrpc(destination_id=request.POST['app_id'], user_id=request.user.id)
+        if not (lwapp and lwdevice):
+            return HttpResponseNotFound()
+        lwdevice.info['destination_id'] = lwapp.info['destination_id']
+        lwdevice.info['destination_type'] = lwapp.info['destination_type']
         lwdevice.save()
     return redirect('csn_devices')
 
@@ -82,16 +85,16 @@ def change_device_app(request):
 @login_required
 def applications(request):
     return render(request, 'csn/applications.html', {
-        'applications': LWApplication.objects.filter(user=request.user)
+        'applications': Destination.get_all_everynet_jsonrpc(user_id=request.user.id)
     })
 
 
 @login_required
 def application(request, app_id):
-    lwapp = get_object_or_404(LWApplication, user=request.user, id=app_id)
+    lwapp = Destination.get_everynet_jsonrpc_with_sensors(destination_id=app_id, user_id=request.user.id)
     return render(request, 'csn/application.html', {
         'application': lwapp,
-    })
+    }) if lwapp else HttpResponseNotFound()
 
 
 @login_required
@@ -100,8 +103,7 @@ def new_app(request):
     if request.method == "POST":
         lwapplication_form = LWApplicationForm(request.POST, user=request.user)
         if lwapplication_form.is_valid():
-            lwapplication = lwapplication_form.save(commit=False)
-            lwapplication.save()
+            Destination.insert_everynet_jsonrpc(lwapplication_form.cleaned_data)
             return redirect('csn_applications')
     return render(request, 'csn/new_application.html', {
         'form': lwapplication_form,
@@ -111,8 +113,8 @@ def new_app(request):
 @login_required
 def delete_app(request):
     if request.method == "POST":
-        lwapp = get_object_or_404(LWApplication, user=request.user, id=request.POST['app_id'])
-        lwapp.delete()
+        if not Destination.delete_everynet_jsonrpc(user_id=request.user.id, destination_id=request.POST['app_id']):
+            return HttpResponseNotFound()
     return redirect('csn_applications')
